feat(marketplace): controller — дробное количество сквозняком (Эпик 17 Story 17.3)
Завершение controller-слоя под asset-количество:
- Все 5 действий цепи (createorder/stockorder/signiss2/signchair/submretrn)
  строят количество через toQuantityAsset(qty, unit) — asset-строка с символом
  единицы; единица денормализована на заказ (marketplace_order.unit_of_measure,
  как price_per_unit) и протянута через create-input/adapter/mapper/domain.
- Колонки количества integer→numeric(18,3) с numericQuantityTransformer:
  order.quantity, offer.quantity_available/blocked/consumed. Расширение на
  synchronize:true → без файла миграции (стенд пересоздаётся).
- GraphQL DTO: поля количества Int→Float (@IsInt→@IsNumber) в issuance/stock/
  order/offer/cart/checkout/return-claim/apl-reception/inventory/event/
  consolidated/offer-input — «Int cannot represent non-integer value: 0.5»
  больше не возникает.
- Number.isInteger снят в order-create/stock/stock-proposal (сайт исходного
  бага) → assertValidQuantity (штука целочисленна, вес/объём дробные).
- Стоимость price×qty остаётся корректной для дробного количества.

Локальный tsc контроллера — чисто (0 ошибок).

+++ components/controller/src/extensions/marketplace/application/dto/marketplace-cart.dto.ts
@@ -1,4 +1,4 @@
-import { Field, Int, ObjectType } from '@nestjs/graphql';
+import { Field, Float, Int, ObjectType } from '@nestjs/graphql';
 
 /**
  * Эпик 16: позиция корзины с обогащением для UI. Реквизиты товара
@@ -13,7 +13,7 @@ export class MarketplaceCartItemDTO {
   @Field(() => String, { description: 'Идентификатор предложения.' })
   public readonly offer_id!: string;
 
-  @Field(() => Int, { description: 'Количество единиц в корзине.' })
+  @Field(() => Float, { description: 'Количество единиц в корзине.' })
   public readonly quantity!: number;
 
   @Field(() => String, {
@@ -61,7 +61,7 @@ export class MarketplaceCartItemDTO {
   })
   public readonly available_on_current_ku!: boolean;
 
-  @Field(() => Int, {
+  @Field(() => Float, {
     nullable: true,
     description:
       'Максимально доступное количество единиц по предложению. null — без ограничения (можно заказать любое количество).',
@@ -96,7 +96,7 @@ export class MarketplaceCartDTO {
   @Field(() => Int, { description: 'Количество разных позиций (строк) в корзине.' })
   public readonly positions_count!: number;
 
-  @Field(() => Int, { description: 'Суммарное количество единиц всех позиций.' })
+  @Field(() => Float, { description: 'Суммарное количество единиц всех позиций.' })
   public readonly total_quantity!: number;
 
   @Field(() => String, { description: 'Итоговая сумма корзины (по доступным к доставке позициям).' })
